Move structured data blocks integration to src, add integration that adds all Yoast block categories at once.
